+++ NEWS.md
@@ -5,6 +5,7 @@
 - `load\_tag` is aware of added columns to status 'Tilt', 'S11', 'ReleaseWetDry', and 'ReleaseTemperature'. If they aren't there they are added with NAs for forward compatability and merge-ability.
 - changed various typos
 - added a bool param to `batch\_load\_tags`, `load\_gonio\_output`, and `load\_tag` to retain original date formats
+- added support for a rare-ish fastgps format in `load\_tag`. Don't worry this format makes even less sense than the other (also) broken format.
 ## MAJOR CHANGES
 - renamed the main branch to 'main'